backend/pyproject.toml
[project]
name = "dailypages-backend"
version = "0.2.0"
description = "DailyPages — FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30",
  "psycopg2-binary>=2.9",
  "alembic>=1.14",
  "pgvector>=0.3.6",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.28",
  "boto3>=1.35",
  "python-multipart>=0.0.20",
  # Parsing pipeline (Phase 4)
  "pypdf>=5.1",
  "pdfminer.six>=20240706",
  "EbookLib>=0.18",
  "beautifulsoup4>=4.12",
  "lxml>=5.3",
  # Async jobs
  "celery>=5.4",
  "redis>=5.2",
  # AI (Phase 5)
  "anthropic>=0.40",
  "voyageai>=0.3",
  "langgraph>=0.2",
  "langchain-core>=0.3",
  "tiktoken>=0.8",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8",
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "scripts*", "tests*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
